backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
redis==5.0.1
drf-yasg==1.21.11
